refactor(debug): split SpDevPanel tabs and styles into separate files
